+++ pkg/uuid/uuid_generator.go
@@ -3,5 +3,5 @@ package uuid
 // UuidGenerator is common uuid generator interface
 type UuidGenerator interface {
 	GenerateUuid(uuidType UuidType) int64
-	GenerateUuids(uuidType UuidType, count uint8) []int64
+	GenerateUuids(uuidType UuidType, count uint16) []int64
 }
